Added working C++ MeshCurver module
The following module works and with good performances. As its name implies, it will 'curve' a given mesh (user chosen) along a Curve3d, that the user can move at will, moving the mesh at the same time. As well as that, the possibility to create a Trimesh Collision Sibling is given, allowing the mesh to be used to create fancy obstacles/environments etc.